+++ src/data/flexboxProperties.ts
@@ -1,4 +1,4 @@
-import { theme } from '../theme/theme';
+import { theme } from '../themes/theme';
 
 export type FlexboxPropertyStyle = {
     parent: { [key: string]: string } | null;
